Alert changes.

+++ src/MatchApi.Application/Features/Teams/Commands/DeleteTeam/DeleteTeamCommandHandler.cs
@@ -1,9 +1,11 @@
 ﻿using MatchApi.Application.Common.Interfaces;
+using MatchApi.Domain.DTOs;
 using MediatR;
+using System.Net;
 
 namespace MatchApi.Application.Features.Teams.Commands.DeleteTeam
 {
-    public class DeleteTeamCommandHandler : IRequestHandler<DeleteTeamCommand>
+    public class DeleteTeamCommandHandler : IRequestHandler<DeleteTeamCommand, ResponseResult<bool>>
     {
         private readonly ITeamRepository _repository;
 
@@ -12,11 +14,23 @@ namespace MatchApi.Application.Features.Teams.Commands.DeleteTeam
             _repository = repository;
         }
 
-        public async Task Handle(
+        public async Task<ResponseResult<bool>> Handle(
             DeleteTeamCommand request,
             CancellationToken cancellationToken)
         {
-            await _repository.DeleteTeamAsync(request.TeamId, cancellationToken);
+            try
+            {
+                var team = await _repository.GetByIdAsync(request.TeamId, cancellationToken);
+                if(team == null)
+                    return new ResponseResult<bool> { Success = false, Message = "Team does not exist." };
+
+                await _repository.DeleteTeamAsync(request.TeamId, cancellationToken);
+                return new ResponseResult<bool> { Success = true, Message = "Team deleted successfully." };
+            }
+            catch (Exception ex)
+            {
+                return new ResponseResult<bool> { Success = false, Error = new Error { Message = ex.Message, StatusCode = HttpStatusCode.InternalServerError.ToString() } };
+            }
         }
     }
 }
